requirements.txt
fastapi==0.104.1
pydantic==2.4.2
uvicorn==0.24.0
openai==1.84.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp==3.9.1
Pillow>=10.0.0
firebase-admin==6.4.0
google-cloud-storage==2.14.0
